feat: extract dashboard module
