+++ dotcraft.sln
@@ -51,6 +51,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotCraft.Sdk", "sdk\dotnet\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Oratorio.Server.Tests", "tests\Oratorio.Server.Tests\Oratorio.Server.Tests.csproj", "{F20FDC5B-1A60-413A-A268-035806F3EEA6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotCraft.Dashboard", "src\DotCraft.Dashboard\DotCraft.Dashboard.csproj", "{FF144851-3FDF-4BF3-B78D-D66F2511E952}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotCraft.Dashboard.Tests", "tests\DotCraft.Dashboard.Tests\DotCraft.Dashboard.Tests.csproj", "{491C57DA-45DF-43C2-AD48-B13644B41292}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -349,6 +353,30 @@ Global
 		{BF62EC8A-1BE3-4D87-87BB-5E05381F660E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{BF62EC8A-1BE3-4D87-87BB-5E05381F660E}.Release|x86.ActiveCfg = Release|Any CPU
 		{BF62EC8A-1BE3-4D87-87BB-5E05381F660E}.Release|x86.Build.0 = Release|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Debug|x64.Build.0 = Debug|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Debug|x86.Build.0 = Debug|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Release|x64.ActiveCfg = Release|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Release|x64.Build.0 = Release|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Release|x86.ActiveCfg = Release|Any CPU
+		{FF144851-3FDF-4BF3-B78D-D66F2511E952}.Release|x86.Build.0 = Release|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Debug|x64.Build.0 = Debug|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Debug|x86.Build.0 = Debug|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Release|x64.ActiveCfg = Release|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Release|x64.Build.0 = Release|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Release|Any CPU.Build.0 = Release|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Release|x86.ActiveCfg = Release|Any CPU
+		{491C57DA-45DF-43C2-AD48-B13644B41292}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
