+++ specs/README.md
@@ -11,9 +11,10 @@ bindings.
 2. [Context Compaction](architecture/context-compaction.md) for compaction backend selection, replacement domains, and recovery.
 3. [Tool Architecture](architecture/tools-architecture.md) for tool identity, authority, execution, result audiences, and presentation boundaries.
 4. [AppServer Protocol](protocols/appserver-protocol.md) for the JSON-RPC projection used by clients and SDKs.
-5. [Plugin Architecture](architecture/plugin-architecture.md), [Skill 2.0](architecture/skill-2.0.md), and [Plugin Registry](architecture/plugin-registry.md) for extension loading, skill resolution, plugin boundaries, and public plugin distribution.
-6. [Desktop Client](clients/desktop-client.md) and [Desktop DESIGN.md](architecture/DESIGN.md) for first-party client behavior and shared Desktop visual rules.
-7. [SDK](sdk/sdk.md) for cross-language client binding expectations and [AppServer Protocol Contracts and SDK Generation](sdk/protocol-contract-generation.md) for the executable wire contract and generated low-level bindings.
+5. [Runtime Module Boundaries](architecture/runtime-module-boundaries.md) for assembly ownership and dependency direction.
+6. [Plugin Architecture](architecture/plugin-architecture.md), [Skill 2.0](architecture/skill-2.0.md), and [Plugin Registry](architecture/plugin-registry.md) for extension loading, skill resolution, plugin boundaries, and public plugin distribution.
+7. [Desktop Client](clients/desktop-client.md) and [Desktop DESIGN.md](architecture/DESIGN.md) for first-party client behavior and shared Desktop visual rules.
+8. [SDK](sdk/sdk.md) for cross-language client binding expectations and [AppServer Protocol Contracts and SDK Generation](sdk/protocol-contract-generation.md) for the executable wire contract and generated low-level bindings.
 
 ## Categories
 
@@ -27,7 +28,7 @@ bindings.
 
 ## Files By Category
 
-- Architecture: [session-core.md](architecture/session-core.md), [context-compaction.md](architecture/context-compaction.md), [tools-architecture.md](architecture/tools-architecture.md), [prompt-composition.md](architecture/prompt-composition.md), [prompt-cache.md](architecture/prompt-cache.md), [responses-provider-history.md](architecture/responses-provider-history.md), [model-runtime.md](architecture/model-runtime.md), [hub-architecture.md](architecture/hub-architecture.md), [plugin-architecture.md](architecture/plugin-architecture.md), [plugin-registry.md](architecture/plugin-registry.md), [skill-2.0.md](architecture/skill-2.0.md), [lsp-plugin.md](architecture/lsp-plugin.md), [DESIGN.md](architecture/DESIGN.md), [openai-subscription-auth.md](architecture/openai-subscription-auth.md)
+- Architecture: [session-core.md](architecture/session-core.md), [runtime-module-boundaries.md](architecture/runtime-module-boundaries.md), [context-compaction.md](architecture/context-compaction.md), [tools-architecture.md](architecture/tools-architecture.md), [prompt-composition.md](architecture/prompt-composition.md), [prompt-cache.md](architecture/prompt-cache.md), [responses-provider-history.md](architecture/responses-provider-history.md), [model-runtime.md](architecture/model-runtime.md), [hub-architecture.md](architecture/hub-architecture.md), [plugin-architecture.md](architecture/plugin-architecture.md), [plugin-registry.md](architecture/plugin-registry.md), [skill-2.0.md](architecture/skill-2.0.md), [lsp-plugin.md](architecture/lsp-plugin.md), [DESIGN.md](architecture/DESIGN.md), [openai-subscription-auth.md](architecture/openai-subscription-auth.md)
 - Features: [subagents.md](features/subagents.md), [agent-profiles.md](features/agent-profiles.md), [agent-teams.md](features/agent-teams.md), [dynamic-workflows.md](features/dynamic-workflows.md), [external-cli-subagent.md](features/external-cli-subagent.md), [goal.md](features/goal.md), [memory-consolidation.md](features/memory-consolidation.md), [dreams.md](features/dreams.md), [lifecycle-hooks.md](features/lifecycle-hooks.md), [context-export-cli.md](features/context-export-cli.md), [automations-lifecycle.md](features/automations-lifecycle.md), [default-chat-workspace.md](features/default-chat-workspace.md), [model-options.md](features/model-options.md), [multi-folder-projects.md](features/multi-folder-projects.md), [remote-server-management.md](features/remote-server-management.md), [chrome-browser-runtime.md](features/chrome-browser-runtime.md), [desktop-inapp-browser.md](features/desktop-inapp-browser.md)
 - Protocols: [appserver-protocol.md](protocols/appserver-protocol.md), [external-channel-adapter.md](protocols/external-channel-adapter.md), [app-binding.md](protocols/app-binding.md)
 - Clients: [desktop-client.md](clients/desktop-client.md)
