+++ src/DotCraft.App/DotCraft.App.csproj
@@ -49,6 +49,7 @@
   <ItemGroup>
     <ProjectReference Include="..\DotCraft.Core\DotCraft.Core.csproj" />
     <ProjectReference Include="..\DotCraft.ContextExport\DotCraft.ContextExport.csproj" />
+    <ProjectReference Include="..\DotCraft.Dashboard\DotCraft.Dashboard.csproj" />
     <ProjectReference Include="..\DotCraft.Agents.OpenAI\DotCraft.Agents.OpenAI.csproj" />
     <ProjectReference Include="..\DotCraft.Agents.Anthropic\DotCraft.Agents.Anthropic.csproj" />
     <ProjectReference Include="..\DotCraft.Automations\DotCraft.Automations.csproj" />
