+++ src/DotCraft.Core/DotCraft.Core.csproj
@@ -9,8 +9,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="Resources\DashBoard.html" />
-    <EmbeddedResource Include="Resources\DashBoardLogin.html" />
     <EmbeddedResource Include="Resources\models.json" />
     <EmbeddedResource Include="Skills\BuiltIn\**\*" />
     <EmbeddedResource Include="Commands\Custom\BuiltIn\**\*" />
