+++ src/DotCraft.Core/Hosting/IOrchestratorSnapshotProvider.cs
@@ -1,20 +1,17 @@
-namespace DotCraft.DashBoard;
+namespace DotCraft.Hosting;
 
 /// <summary>
-/// Allows orchestrator modules to expose their runtime state to the dashboard API.
-/// Implement this interface in any orchestrator that should be queryable via
-/// <c>/dashboard/api/orchestrators/{name}/state</c> and <c>/refresh</c>.
+/// Allows orchestrator modules to expose runtime state to host-provided observability surfaces.
 /// </summary>
 public interface IOrchestratorSnapshotProvider
 {
     /// <summary>
-    /// Unique channel/module name used as the URL segment for the dashboard endpoints.
+    /// Unique module name used to identify the snapshot.
     /// </summary>
     string Name { get; }
 
     /// <summary>
     /// Returns a serializable snapshot of the current orchestrator state.
-    /// The returned object is serialized to JSON by the dashboard middleware.
     /// </summary>
     object GetSnapshot();
