+++ src/DotCraft.Dashboard/DashBoard/DashBoardThreadOperationStore.cs
