+++ src/DotCraft.Dashboard/DashBoard/IDashBoardSessionHandler.cs
