+++ src/DotCraft.Dashboard/Resources/DashBoard.html
