+++ src/DotCraft.Dashboard/Resources/DashBoardLogin.html
