+++ tests/DotCraft.Dashboard.Tests/DotCraft.Dashboard.Tests.csproj
@@ -1,0 +1,24 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <IsPackable>false</IsPackable>
+    <AssemblyName>DotCraft.Dashboard.Tests</AssemblyName>
+    <RootNamespace>DotCraft.Dashboard.Tests</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="coverlet.collector" Version="6.0.4" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.14.1" />
+    <PackageReference Include="xunit" Version="2.9.3" />
+    <PackageReference Include="xunit.runner.visualstudio" Version="3.1.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\DotCraft.Dashboard\DotCraft.Dashboard.csproj" />
+    <ProjectReference Include="..\..\src\DotCraft.Core\DotCraft.Core.csproj" />
+  </ItemGroup>
+
+</Project>
